fix: Entity PK 타입 Integer 통일 및 양방향 JSON 순환 차단

+++ backend/src/main/java/com/smartfridge/repository/CookHistoryRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CookHistoryRepository extends JpaRepository<CookHistory, Long> {
+public interface CookHistoryRepository extends JpaRepository<CookHistory, Integer> {
 
     List<CookHistory> findAllByOrderByCookedAtDesc();
 }
